Conflicts:
	source/blender/blenkernel/intern/DerivedMesh.c
	source/blender/blenkernel/intern/rigidbody.c
	source/blender/blenkernel/intern/smoke.c
	source/blender/editors/space_buttons/buttons_context.c

+++ source/blender/render/intern/source/sunsky.c
@@ -28,7 +28,6 @@
 
 #include "sunsky.h"
 #include "BLI_math.h"
-#include "BKE_global.h"
 
 /**
  * These macros are defined for vector operations
@@ -109,8 +108,8 @@ static float AngleBetween(float thetav, float phiv, float theta, float phi)
  * */
 static void DirectionToThetaPhi(float *toSun, float *theta, float *phi)
 {
-	*theta = acos(toSun[2]);
-	if (fabs(*theta) < 1e-5)
+	*theta = acosf(toSun[2]);
+	if (fabsf(*theta) < 1e-5f)
 		*phi = 0;
 	else
 		*phi = atan2(toSun[1], toSun[0]);
